fix build.

+++ tests/test_value_array.cc
@@ -10,16 +10,16 @@ using namespace tora;
 
 int main() {
     SharedPtr<ArrayValue> av = new ArrayValue();
-    is(av->size(), 0UL);
+    is(av->size(), (size_t) 0UL);
     av->push_back(new IntValue(49));
-    is(av->size(), 1UL);
+    is(av->size(), (size_t) 1UL);
     SharedPtr<Value> v = new IntValue(55);
     av->set_item(1, v);
     is(av->at(1)->upcast<IntValue>()->int_value(), 55);
     av->set_item(1, new IntValue(88));
     is(av->at(1)->upcast<IntValue>()->int_value(), 88);
     // is((void*)v.get(), (void*)av->at(1).get());
-    is(av->size(), 2UL);
+    is(av->size(), (size_t) 2UL);
 
     done_testing();
 }
